refactor(sidebar): déplacer les composants vers la structure Atomic Design globale
- atoms/ molecules/ organisms/ au niveau de src/components/ (et non plus dans layout/sidebar/)
- Mettre à jour tous les imports cross-couches vers des chemins absolus (@/components/...)
- Mettre à jour header.tsx et app/(dashboard)/layout.tsx
- Déplacer les tests en miroir (atoms/ molecules/ organisms/)
- Documenter l'architecture frontend Atomic Design dans docs/ARCHITECTURE.md
