+++ client/src/app/(dashboard)/layout.tsx
@@ -1,5 +1,5 @@
 import { Header } from "@/components/layout/header";
-import { Sidebar } from "@/components/layout/sidebar";
+import { Sidebar } from "@/components/organisms/sidebar";
 
 export default function DashboardLayout({
   children,
