+++ client/src/components/atoms/sidebar/sidebar-logo.tsx
