+++ client/src/components/molecules/sidebar/desktop-project-item.tsx
