+++ client/src/components/molecules/sidebar/mobile-project-item.tsx
