+++ client/src/components/organisms/sidebar/mobile-sidebar.tsx
@@ -1,14 +1,13 @@
 "use client";
 
 import { useTranslations } from "next-intl";
-import { SidebarLogo } from "./atoms/sidebar-logo";
-import { SidebarNav } from "./organisms/sidebar-nav";
-import { ProjectsSection } from "./organisms/projects-section";
-import { OrganizationSection } from "./organisms/organization-section";
-import { UserMenu } from "./organisms/user-menu";
+import { SidebarLogo } from "@/components/atoms/sidebar/sidebar-logo";
+import { SidebarNav } from "./sidebar-nav";
+import { ProjectsSection } from "./projects-section";
+import { OrganizationSection } from "./organization-section";
 import { useSidebarData } from "./use-sidebar-data";
 
-export function Sidebar() {
+export function MobileSidebar() {
   const t = useTranslations("sidebar");
   const {
     personalProjects,
@@ -19,12 +18,12 @@ export function Sidebar() {
   } = useSidebarData();
 
   return (
-    <aside className="hidden h-full w-64 border-r bg-white dark:bg-muted/30 md:flex md:flex-col">
+    <div>
       <SidebarLogo />
-      <nav className="flex-1 space-y-1 overflow-y-auto p-3">
-        <SidebarNav showIcons />
+      <nav className="space-y-1 p-4">
+        <SidebarNav matchPrefix />
         <ProjectsSection
-          variant="desktop"
+          variant="mobile"
           title={t("myProjects")}
           projects={personalProjects}
           isLoading={isLoadingProjects}
@@ -35,14 +34,13 @@ export function Sidebar() {
         {sortedOrganizations.map((org) => (
           <OrganizationSection
             key={org.id}
-            variant="desktop"
+            variant="mobile"
             organization={org}
             projects={organizationProjectsMap.get(org.id) ?? []}
             canCreate={editableOrganizationIds.has(org.id)}
           />
         ))}
       </nav>
-      <UserMenu />
-    </aside>
+    </div>
   );
 }
