+++ client/src/components/organisms/sidebar/projects-section.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import { useTranslations } from "next-intl";
-import { SidebarSectionHeader } from "../atoms/sidebar-section-header";
-import { DesktopProjectItem } from "../molecules/desktop-project-item";
-import { MobileProjectItem } from "../molecules/mobile-project-item";
+import { SidebarSectionHeader } from "@/components/atoms/sidebar/sidebar-section-header";
+import { DesktopProjectItem } from "@/components/molecules/sidebar/desktop-project-item";
+import { MobileProjectItem } from "@/components/molecules/sidebar/mobile-project-item";
 import type { ProjectResponse } from "@/lib/types/api";
 
 interface ProjectsSectionProps {
